patch 4.2.52.4
Fix a bug where recovery of a page split after a non-transactional update
to the next page would fail to update the back pointer. This bug could
result in log sequence errors and missing data when when the database is
walked with a cursor, after data was initially loaded into a database
non-transactionally (for example, by using the db_load utility).